Simplify find_check_items a bit more
As it turns out, checkbox ids always have a format of `check_xxx`, where xxx is a compressed id or index. xxx never contains underscores.

And even if it did, then the output of find_checked_items would be an underscore-separated string of integers (e.g. "10_12_13"), which would inevitably led to bugs in places where we pass this value to `find()` or `where()` calls, because AR would cast such things to_i, and it would return 101213, which clearly is not what we want.